+++ src/main/java/roomescape/core/controller/ThemeController.java
@@ -1,17 +1,10 @@
 package roomescape.core.controller;
 
-import jakarta.validation.Valid;
-import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.core.dto.theme.ThemeRequest;
 import roomescape.core.dto.theme.ThemeResponse;
 import roomescape.core.service.ThemeService;
 
@@ -24,13 +17,6 @@ public class ThemeController {
         this.themeService = themeService;
     }
 
-    @PostMapping
-    public ResponseEntity<ThemeResponse> create(@Valid @RequestBody final ThemeRequest request) {
-        final ThemeResponse response = themeService.create(request);
-        return ResponseEntity.created(URI.create("/themes/" + response.getId()))
-                .body(response);
-    }
-
     @GetMapping
     public ResponseEntity<List<ThemeResponse>> findAll() {
         return ResponseEntity.ok(themeService.findAll());
@@ -39,11 +25,5 @@ public class ThemeController {
     @GetMapping("/popular")
     public ResponseEntity<List<ThemeResponse>> findPopular() {
         return ResponseEntity.ok(themeService.findPopularTheme());
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable("id") final long id) {
-        themeService.delete(id);
-        return ResponseEntity.noContent().build();
     }
 }
