+++ src/test/java/roomescape/core/controller/AdminControllerTest.java
@@ -7,6 +7,9 @@ import java.time.format.DateTimeFormatter;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullAndEmptySource;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.MediaType;
@@ -14,6 +17,8 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import roomescape.core.dto.auth.TokenRequest;
 import roomescape.core.dto.reservation.ReservationRequest;
+import roomescape.core.dto.reservationtime.ReservationTimeRequest;
+import roomescape.core.dto.theme.ThemeRequest;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
@@ -62,6 +67,84 @@ class AdminControllerTest {
     }
 
     @Test
+    @DisplayName("시간 관리 페이지로 이동한다.")
+    void moveToTimeManagePage() {
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .when().get("/admin/time")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @ParameterizedTest
+    @NullAndEmptySource
+    @ValueSource(strings = {" ", "10:89"})
+    @DisplayName("시간 생성 시, startAt 값의 형식이 올바르지 않으면 예외가 발생한다.")
+    void validateTimeCreateWithEmpty(final String startAt) {
+        ReservationTimeRequest request = new ReservationTimeRequest(startAt);
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("시간 생성 시, startAt 값이 중복되면 예외가 발생한다.")
+    void validateTimeDuplicated() {
+        ReservationTimeRequest request = new ReservationTimeRequest("10:00");
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(201);
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("시간 삭제 시, 해당 시간을 참조하는 예약이 있으면 예외가 발생한다.")
+    void validateTimeDelete() {
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .when().delete("/admin/times/1")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("시간 삭제 시, 해당 시간을 참조하는 예약이 없으면 삭제된다.")
+    void deleteTime() {
+        ReservationTimeRequest request = new ReservationTimeRequest("10:00");
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/times")
+                .then().log().all()
+                .statusCode(201);
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .when().delete("/admin/times/4")
+                .then().log().all()
+                .statusCode(204);
+    }
+
+    @Test
     @DisplayName("테마 관리 페이지로 이동한다.")
     void moveToThemeManagePage() {
         RestAssured.given().log().all()
@@ -71,14 +154,97 @@ class AdminControllerTest {
                 .statusCode(200);
     }
 
-    @Test
-    @DisplayName("시간 관리 페이지로 이동한다.")
-    void moveToTimeManagePage() {
+    @ParameterizedTest
+    @NullAndEmptySource
+    @ValueSource(strings = {" "})
+    @DisplayName("테마 생성 시, name 값이 올바르지 않으면 예외가 발생한다.")
+    void validateThemeWithNameEmpty(final String name) {
+        ThemeRequest request = new ThemeRequest(name, "우테코 레벨2를 탈출하는 내용입니다.",
+                "https://i.pinimg.com/236x/6e/bc/46/6ebc461a94a49f9ea3b8bbe2204145d4.jpg");
+
         RestAssured.given().log().all()
                 .cookies("token", accessToken)
-                .when().get("/admin/time")
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/themes")
                 .then().log().all()
-                .statusCode(200);
+                .statusCode(400);
+    }
+
+    @ParameterizedTest
+    @NullAndEmptySource
+    @ValueSource(strings = {" "})
+    @DisplayName("테마 생성 시, description 값이 올바르지 않으면 예외가 발생한다.")
+    void validateThemeWithDescriptionEmpty(final String description) {
+        ThemeRequest request = new ThemeRequest("레벨2 탈출", description,
+                "https://i.pinimg.com/236x/6e/bc/46/6ebc461a94a49f9ea3b8bbe2204145d4.jpg");
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @ParameterizedTest
+    @NullAndEmptySource
+    @ValueSource(strings = {" "})
+    @DisplayName("테마 생성 시, thumbnail 값이 올바르지 않으면 예외가 발생한다.")
+    void validateThemeWithThumbnailEmpty(final String thumbnail) {
+        ThemeRequest request = new ThemeRequest("레벨2 탈출", "우테코 레벨2를 탈출하는 내용입니다.", thumbnail);
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("테마 생성 시, name 값이 중복이면 예외가 발생한다.")
+    void validateThemeWithDuplicatedName() {
+        ThemeRequest request = new ThemeRequest("레벨2 탈출", "우테코 레벨2를 탈출하는 내용입니다.",
+                "https://i.pinimg.com/236x/6e/bc/46/6ebc461a94a49f9ea3b8bbe2204145d4.jpg");
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(201);
+
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .contentType(ContentType.JSON)
+                .body(request)
+                .when().post("/admin/themes")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("테마 삭제 시, 해당 테마를 참조하는 예약이 있으면 예외가 발생한다.")
+    void validateThemeDelete() {
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .when().delete("/admin/themes/1")
+                .then().log().all()
+                .statusCode(400);
+    }
+
+    @Test
+    @DisplayName("테마 삭제 시, 해당 테마를 참조하는 예약이 없으면 테마가 삭제된다.")
+    void deleteTheme() {
+        RestAssured.given().log().all()
+                .cookies("token", accessToken)
+                .when().delete("/admin/themes/2")
+                .then().log().all()
+                .statusCode(204);
     }
 
     @Test
